# docs/swagger.yaml
basePath: /api/
definitions:
  serialization.Response:
    additionalProperties: true
    type: object
host: http://espresso.nctu.me:8080/
info:
  contact:
    email: davidleitw@gmail.com
    name: davidleitw
  description: Expreesso calendar的api文檔, 方便串接前後端
  license:
    name: Apache 2.0
  title: Espresso Example API File
  version: 0.0.1
paths:
  /api/calendar/{ID}/createNewEvent:
    post:
      consumes:
      - application/json
      parameters:
      - description: 填入事件標題
        in: header
        name: Title
        required: true
        type: string
      - description: 填入事件開始時間
        in: header
        name: StartTime
        required: true
        type: string
      - description: 填入事件的結束時間
        in: header
        name: EndTime
        required: true
        type: string
      - description: 填入事件想要提前幾分鐘提醒 以-3h這種形式
        in: header
        name: RemindTime
        required: true
        type: string
      - description: 填入事件內容
        in: header
        name: Context
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
        "400": {}
        "500": {}
      summary: 創建事件
      tags:
      - Calendar
  /api/calendar/{ID}/deleteEvent:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 填入想要刪除事件的標題
        in: header
        name: title
        required: true
        type: string
      - description: 填入想要刪除事件的開始時間
        in: header
        name: start_time
        required: true
        type: string
      - description: 填入想要刪除事件需要提前幾分鐘提醒 以-3h這種形式填入(創建事件要填的那個寫法)
        in: header
        name: remind_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204": {}
        "500": {}
      summary: 刪除指定的事件
      tags:
      - Calendar
  /api/calendar/{ID}/getAllEvent:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serialization.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/serialization.Response'
      summary: 登入時該用戶所有資料的概要
      tags:
      - Calendar
  /api/calendar/{ID}/getEventInfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 填入想要查詢事件的標題
        in: header
        name: title
        required: true
        type: string
      - description: 填入想要查詢事件的開始時間
        in: header
        name: start_time
        required: true
        type: string
      - description: 填入想要查詢事件需要提前幾分鐘提醒 以-3h這種形式填入(創建事件要填的那個寫法)
        in: header
        name: remind_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
        "500": {}
      summary: 獲得某個特定事件的資料
      tags:
      - Calendar
  /api/calendar/{ID}/updateEvent:
    put:
      consumes:
      - application/json
      parameters:
      - description: 填入修改前的事件標題
        in: header
        name: OldTitle
        required: true
        type: string
      - description: 填入修改前的開始時間
        in: header
        name: OldStart
        required: true
        type: string
      - description: 填入修改前的欲提醒時間
        in: header
        name: OldRemind
        required: true
        type: string
      - description: 填入修改後的新標題
        in: header
        name: Title
        required: true
        type: string
      - description: 填入修改後的新開始時間
        in: header
        name: StartTime
        required: true
        type: string
      - description: 填入修改後的新結束時間
        in: header
        name: EndTime
        required: true
        type: string
      - description: 填入修改後的新提醒時間 以-3h這種形式
        in: header
        name: RemindTime
        required: true
        type: string
      - description: 填入修改後的事件內容
        in: header
        name: Context
        required: true
        type: string
      - description: 填入修改後的參考網址
        in: header
        name: ReferenceUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204": {}
        "500": {}
      summary: 更新某個事件的內容
      tags:
      - Calendar
  /api/user/userLogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用戶電子信箱(帳號)
        in: header
        name: account
        required: true
        type: string
      - description: 用戶密碼
        in: header
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/serialization.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/serialization.Response'
      summary: 用戶登入
      tags:
      - User
  /api/user/userLogout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/serialization.Response'
      summary: 用戶登出
      tags:
      - User
  /api/user/userRegister:
    post:
      consumes:
      - application/json
      parameters:
      - description: 填入用戶的電子信箱
        in: header
        name: UserMail
        required: true
        type: string
      - description: 填入用戶的使用者名稱 可隨時更改
        in: header
        name: UserName
        required: true
        type: string
      - description: 填入用戶的密碼 需要八碼以上
        in: header
        name: UserPass
        required: true
        type: string
      - description: 密碼確認, 需要UserPass參數相同
        in: header
        name: UserPassConfirm
        required: true
        type: string
      - description: 用戶的電話號碼 如果使用者沒填 前端會補上00000000
        in: header
        name: USerRid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serialization.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/serialization.Response'
      summary: 用戶註冊
      tags:
      - User
  /api/user/usercheckLogin:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: "false"
          schema:
            type: string
      summary: 確認登入狀態
      tags:
      - User
swagger: "2.0"